debugging code
IO.java
since we dont have a trigger yet, we made the method that checks if the trigger is activated to always return true.

Drive.java
with normal DigitalInput sensors, there seems to be some sort of error where it says "HAL resource already allocated", making it so that the code can't instantiate.

So, we've just set the sensors to booleans and manually set them to true or false, and then changed the sensor2.get() to sensor2 since it's now a boolean

Co-Authored-By: sci-phy <sci-phy@users.noreply.github.com>
Co-Authored-By: chanduniverse <chanduniverse@users.noreply.github.com>
Co-Authored-By: Ravi Grewal <rgcoderx@users.noreply.github.com>