Jevois Ball Detection
Will bring the ball and jevois to the next meeting. Currently the jevois is able to print out how far the center of the ball is away from the center of the camera but it can only print it out to USB not serial. Will look into this next meeting